% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,PPsegOutput,numeric-method}
\alias{plot,PPsegOutput,numeric-method}
\title{Plot with PPseg}
\usage{
\S4method{plot}{PPsegOutput,numeric}(x, y)
}
\arguments{
\item{x}{: a PPsegOutput object or a PPsegOutputReg object
y : the number of groups displayed ( the values must be in PPsegOutput )}
}
\description{
The function plot make a graph data as a function of a discretization of the interval "range" in the PPsegOutput object.
Moreover, there are on the graph different lines representing the value of the Poisson parameters of each group with their starting time of a section.
}
