% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PPseg-class}
\alias{PPseg-class}
\title{PPseg Class}
\value{
Output elements of EM 
-> nb_group : number of groups
-> log_likelihood
-> parameter : a list with beta and lambda parameters
-> partition : a list with partition estimated by EM ( last step (E) ) and MAP the maximum posteriori of the partition estimated
-> segmentation : a list with segmentation estimated by the logistic part with beta estiamted by EM and MAP the maximum posteriori of the segmentation estimated

   Output elements of criteria : list of different criterias perform
}
\description{
PPseg Class groups all the output elements of EM and the selection criteria values for a certain value of nb_group.
}
