% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppseg.R
\name{ppseg}
\alias{ppseg}
\title{ppseg}
\usage{
ppseg(donnees, gtests, S = 1000, nb_tests = 4, nbcpu = 1)
}
\arguments{
\item{donnees}{is a matrix with the individuals in rows and the time step in column}

\item{cova}{is a array individuals x time step x covariates}

\item{gtests}{is the vector of groups will be used}

\item{S}{is the number of iterations used in the different criteria that used a importance sampling}

\item{nb_tests}{is the number of EM algorithms performed}

\item{nbcpu}{is the number of cores used with a function coming from "parallel"}
}
\value{
a PPsegOutputReg object or a PPSegOutput object if cova is empty
}
\description{
ppseg is a function that performs nb_tests algorithms EM
 and select the results with the minimum likelihood. Then, calculates the different criteria.
 This operation is performed for each number of groups in the vector gtests.
 Algorithm EM is based on a Poisson process mixing model with logistic weights.
}
\examples{
set.seed(123)
x <- c(rpois(30,5),rpois(10,10),rpois(30,15))
out  <- ppseg(donnees = x, g = c(1,2,3,4,5), S = 1000, nb_tests = 6, nbcpu = 3)


}
\author{
Julien Gheysens (email : julien.gheysens.59@gmail.com)
}
