% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PPsegData-class}
\alias{PPsegData-class}
\title{PPsegData Class}
\usage{
PPsegData(data,range=(0,1))
}
\value{
Input elements of EM 
-> Data
-> n : Number of individuals
-> TT : Number of steps of time
-> range : data's range
}
\description{
PPsegData Class groups all input elements of EM.
}
