% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PPsegDataReg-class}
\alias{PPsegDataReg-class}
\title{PPsegDataReg Class}
\usage{
PPsegDataReg(data = data,y = covariates,range = (0,1))
}
\value{
Input elements of EM 
-> Data
-> y : covariates
-> n : Number of individuals
-> TT : Number of steps of time
-> range : data's range
}
\description{
PPsegDataReg Class groups all input elements of EM.
}
