% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PPsegOutput-class}
\alias{PPsegOutput-class}
\title{PPsegOutput Class}
\value{
Input elements of EM : a PPsegData object
   Output elementsof EM : a list with PPseg objects
}
\description{
PPsegOutput Class groups all input and output elements of EM.
}
