% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PPsegOutputReg-class}
\alias{PPsegOutputReg-class}
\title{PPsegOutputReg Class}
\value{
Input elements of EM : a PPsegDataReg object
   Output elementsof EM : a list with PPsegreg objects
}
\description{
PPsegOutput Class groups all input and output elements of EM.
}
