% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppseg.R
\name{selection_EM}
\alias{selection_EM}
\title{Selection EM}
\usage{
selection_EM(donnees, g, nb_tests = 4, nbcpu = 3)
}
\arguments{
\item{donnees}{is a matrix with the individuals in rows and the time step in column}

\item{g}{is the number of groups}

\item{nb_tests}{is the number of EM algorithms performed}

\item{nbcpu}{is the number of cores used with a function coming from "parallel"}
}
\value{
a list
}
\description{
selection_EM is a function that performs nb_tests algorithms EM
 and select the results with the minimum likelihood. Algorithm EM is based on a Poisson process mixing model with logistic weights
}
\examples{
set.seed(123)
x <- c(rpois(30,5),rpois(10,10),rpois(30,15))
out  <- selection_EM(donnees = x, g = 3, nb_tests = 6, nbcpu = 3)



}
\author{
Julien Gheysens (email : julien.gheysens.59@gmail.com)
}
