% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppseg.R
\name{selection_EMreg}
\alias{selection_EMreg}
\title{Selection EMreg}
\usage{
selection_EMreg(donnees, cova = array(0), g, nb_tests = 4, nbcpu = 3)
}
\arguments{
\item{donnees}{is a matrix with the individuals in rows and the time step in column}

\item{cova}{is a array individuals x time step x covariates}

\item{g}{is the number of groups}

\item{nb_tests}{is the number of EM algorithms performed}

\item{nbcpu}{is the number of cores used with a function coming from "parallel"}
}
\value{
a list
}
\description{
selection_EMreg is a function that performs nb_tests algorithms EM
 and select the results with the minimum likelihood.
 Algorithm EM is based on a Poisson process mixing model with logistic weights and Poisson parameters dependent on covariates
}
\author{
Julien Gheysens (email : julien.gheysens.59@gmail.com)
}
