% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary,PPsegOutputReg-method}
\alias{summary,PPsegOutputReg-method}
\title{Summary with PPseg}
\usage{
\S4method{summary}{PPsegOutputReg}(object)
}
\arguments{
\item{object}{: a PPsegOutput object or a PPsegOutputReg object}
}
\description{
The function summary show the number of groups selected by the different criteria with the maximum values used in a PPsegOutput object or PPsegOutput object.
}
